add clustering feature for charger markers
