+++ package.json
@@ -14,6 +14,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-leaflet": "^5.0.0",
+    "react-leaflet-cluster": "^2.1.0",
     "react-router-dom": "^7.7.1"
   },
   "devDependencies": {
