+++ src/assets/charger-station-icon.png
