+++ src/assets/marker-icon-red.png
