+++ src/data/chargerLocations.js
@@ -1,0 +1,17 @@
+const chargerLocations = [
+  { id: 1, lat: -38.1582, lng: 144.3745, name: 'East Geelong Station' },
+  { id: 2, lat: -38.1800, lng: 144.3500, name: 'Belmont Station' },
+  { id: 3, lat: -38.2054, lng: 144.3393, name: 'Grovedale Station' },
+  { id: 4, lat: -38.1407, lng: 144.3367, name: 'Geelong West Station' },
+  { id: 5, lat: -38.1649, lng: 144.3928, name: 'Newcomb Station' },
+  { id: 6, lat: -38.1155, lng: 144.3491, name: 'North Geelong Station' },
+  { id: 7, lat: -38.1760, lng: 144.3176, name: 'Highton Station' },
+  { id: 8, lat: -38.1484, lng: 144.3570, name: 'Geelong CBD Station' },
+  { id: 9, lat: -38.2110, lng: 144.3068, name: 'Waurn Ponds Station' },
+  { id: 10, lat: -38.0730, lng: 144.3591, name: 'Corio Station' },
+  { id: 11, lat: -38.1507, lng: 144.3344, name: 'Corio Station' },
+  { id: 12, lat: -38.2363, lng: 144.3708, name: 'Armstrong Creek Station' },
+  { id: 13, lat: -38.2140, lng: 144.3686, name: 'Charlemont Station' },
+];
+
+export default chargerLocations;
